Security fix
- add AllowEdit() check when update or remove in an Admin program

+++ modules/Eligibility/Activities.php
@@ -23,7 +23,7 @@ if($_REQUEST['month_values'] && $_POST['month_values'])
 	$_POST['values'] = $_REQUEST['values'];
 }
 
-if($_REQUEST['values'] && $_POST['values'])
+if($_REQUEST['values'] && $_POST['values'] && AllowEdit())
 {
 	foreach($_REQUEST['values'] as $id=>$columns)
 	{	
@@ -65,7 +65,7 @@ if($_REQUEST['values'] && $_POST['values'])
 
 DrawHeader(ProgramTitle());
 
-if($_REQUEST['modfunc']=='remove')
+if($_REQUEST['modfunc']=='remove' && AllowEdit())
 {
 	if(DeletePrompt(_('Activity')))
 	{
